fix bugs

+++ front-end/src/tools/Card.jsx
@@ -39,7 +39,7 @@ export function FoodCard({ clsName, information, calledFunction }) {
             <CardContent sx={{ display: 'flex', flexDirection: 'column', width: '100%', height: '100%' }} >
                 <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between' }}>
                     <div style={{ fontSize: 30, margin: 5, fontWeight: 'bold', color: 'rgb(56, 56, 56)', marginLeft: 10 }}>{information.name}</div>
-                    <div style={{ fontSize: 30, margin: 5, fontWeight: 'bold', color: 'rgb(56, 56, 56)', marginRight: 10 }}>${Number.parseInt(information.price)}</div>
+                    <div style={{ fontSize: 30, margin: 5, fontWeight: 'bold', color: 'rgb(56, 56, 56)', marginRight: 10 }}>${Number.parseFloat(information.price)}</div>
                 </div>
                 <div style={{ fontFamily: 'sans-serif', marginTop: 8, fontSize: 15, padding: 7, wordWrap: 'break-word', color: "gray" }}>{information.detail}</div>
             </CardContent>
